`auth`: split out web login to new `webauth` pkg
